changed class files
